Handle case where N="" and S="" for cSHAKE.
When both the customization and function name strings are the empty
string, cSHAKE is equivalent to SHAKE as defined in Section 3.3 of
NIST SP 800-185. Libkeccak doesn't handle this case in its implementation
of cSHAKE, and has a precondition to prohibit using it with both
empty strings. So to handle this is ksum we detect this case and
fall back to the SHAKE implementation ourselves.