Better error message when attempting to open directory.
GNAT's implementation of Ada.Text_IO.Open seems to allow us to open
a directory, which then raises a Device_Error when attempting to read
from the file handle, without any meaningful error message in the
exception.

This change adds explicit detection for directories so that we can
display a better error message.