Refactor ksum.adb into a separate package.
The bulk of the ksum procedure is to perform hashing of files and
write output checksums. To support the checking mode of operation
this functionality is refactored into its own subroutine so that
the ksum procedure now just handles argument parsing and then kicking
off either checksum printing or checking.