Implement file checking.
This implements the behaviour of the --check switch to open the
requested file(s) and then check the checksums of each properly
formatted line.

A non-zero exit code is set if any file has an invalid checksum,
and the appropriate warning messages are displayed in a format
similar to sha256sum.