Fix storage error when reading excessively long lines.
This was caused by calling Ada.Text_IO.Get_Line which returns a String
via the secondary stack. If the line in the file is excessively long
then this can cause the stack to overflow, resulting in a Storage_Error
(which is caught and displayed).

This can occur if a checksum file contains a very long checksum.