Add --version switch
The version information is stored in the file version.ads, which is
generated from the output of git describe --tags. This includes the
tag name (e.g. v0.1.0), the number of commits since the tag, and
the shortened commit hash that is built.

A new script build.sh generates the version file then invokes gprbuild
to build ksum.