Change loop exit condition when reading files.
This changes the way that files are read by avoiding use of
the `End_Of_File` function to detect when the end of file is reached.
This has problems on Windows systems when a SUB byte `16#1A#` is
reached in the input, which `End_Of_File` treats as an EOF.

Instead, we stop reading from the file when no more bytes can be
read from it. I.e. when the `Length` returned is 0.