Fix failure to read past SUB characters in Windows (#8)
* Change loop exit condition when reading files.

This changes the way that files are read by avoiding use of
the `End_Of_File` function to detect when the end of file is reached.
This has problems on Windows systems when a SUB byte `16#1A#` is
reached in the input, which `End_Of_File` treats as an EOF.

Instead, we stop reading from the file when no more bytes can be
read from it. I.e. when the `Length` returned is 0.

* Fix run_tests.py on Windows.

This makes the following changes to handle Windows-specific quirks:
* Handle carriage returns characters in the output from ksum.
* Close temporary files before running ksum, and delete them manually.

* Add a test to check line endings for checksum files.

This test checks that ksum can handle both LF and CRLF line endings
in checksum files.

* Trim CR characters from lines in checksum files.

This is necessary for Unix systems where Get_Line does not consume
the CR characters. Therefore, we need to manually strip them to
allow ksum on Unix to process files generated on Windows.